REFACTORING:: modal 추가를 위한 폴더 정리
