+++ src/layout/CommentModal/index.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import * as S from "./style";
+
+const CommentModal = () => {
+  return (
+    <S.Page>
+      <S.Modal></S.Modal>
+    </S.Page>
+  );
+};
+
+export default CommentModal;
